Move non-test parts of ZEO.tests.forker to ZEO._forker for ZEO.server.
FIxes #105.
